Move database module to the metrics folder
